refactor getRastaurants api, create usecontext
