+++ src/contexts/index.js
@@ -1,3 +1,4 @@
 import React from "react";
 export const WeatherContext = React.createContext({});
 export const LandscapesContext = React.createContext({});
+export const RestaurantsContext = React.createContext({});
